requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==20.2.0             # via pytest
certifi==2020.6.20        # via -r requirements.in, requests, sentry-sdk
chardet==3.0.4            # via -r requirements.in, requests
click==7.1.2              # via -r requirements.in, pip-tools
configobj==5.0.6          # via -r requirements.in, slick-bitcoinrpc
constants==0.6.0
ecdsa==0.16.0             # via -r requirements.in
idna==2.10                # via -r requirements.in, requests
iniconfig==1.1.1          # via pytest
packaging==20.4           # via pytest
pbkdf2==1.3               # via -r requirements.in
pip-tools==5.3.1          # via -r requirements.in
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyaes==1.6.1              # via -r requirements.in
pyblake2==1.1.2           # via -r requirements.in
pycurl==7.43.0.6          # via -r requirements.in, slick-bitcoinrpc
pyparsing==2.4.7          # via packaging
pytest==6.1.1             # via -r requirements.in
python-dotenv==0.14.0     # via -r requirements.in
requests==2.24.0          # via -r requirements.in
sentry-sdk==0.19.4        # via -r requirements.in
six==1.15.0               # via -r requirements.in, configobj, ecdsa, packaging, pip-tools
slick-bitcoinrpc==0.1.4   # via -r requirements.in
toml==0.10.1              # via pytest
ujson==3.2.0              # via -r requirements.in, slick-bitcoinrpc
urllib3==1.25.10          # via -r requirements.in, requests, sentry-sdk

# The following packages are considered to be unsafe in a requirements file:
# pip
